test: merge 2 DataValueSetService integration tests
